requirements.txt
streamlit
openai
llama-index
python-dotenv